requirements.txt
paho-mqtt==2.1.0
requests==2.32.3
websockets==14.1
zeep==4.3.1